+++ private/pkg/protosource/protosource.go
@@ -841,6 +841,43 @@ func PackageToNameToService(files ...File) (map[string]map[string]Service, error
 	return packageToNameToService, nil
 }
 
+// PackageToDirectlyImportedPackageToFileImports maps packages to directly imported packages
+// to the FileImports that import this package.
+//
+// For example, if package a imports package b via c/d.proto and c/e.proto, this will have
+// a -> b -> [c/d.proto, c/e.proto].
+//
+// A directly imported package will not be equal to the package, i.e. there will be no a -> a.
+//
+// Files with no packages are included with key "" to be consistent with other functions.
+func PackageToDirectlyImportedPackageToFileImports(files ...File) (map[string]map[string][]FileImport, error) {
+	filePathToFile, err := FilePathToFile(files...)
+	if err != nil {
+		return nil, err
+	}
+	packageToDirectlyImportedPackageToFileImports := make(map[string]map[string][]FileImport)
+	for _, file := range files {
+		pkg := file.Package()
+		directlyImportedPackageToFileImports, ok := packageToDirectlyImportedPackageToFileImports[pkg]
+		if !ok {
+			directlyImportedPackageToFileImports = make(map[string][]FileImport)
+			packageToDirectlyImportedPackageToFileImports[pkg] = directlyImportedPackageToFileImports
+		}
+		for _, fileImport := range file.FileImports() {
+			if importedFile, ok := filePathToFile[fileImport.Import()]; ok {
+				importedPkg := importedFile.Package()
+				if importedPkg != pkg {
+					directlyImportedPackageToFileImports[importedFile.Package()] = append(
+						directlyImportedPackageToFileImports[importedPkg],
+						fileImport,
+					)
+				}
+			}
+		}
+	}
+	return packageToDirectlyImportedPackageToFileImports, nil
+}
+
 // NameToMethod maps the Methods in the Service to a map from name to Method.
 //
 // Returns error if Methods do not have unique names within the Service, which should